requirements.txt
Pillow
google-cloud
google-cloud-storage